+++ pf_delivery.sh
@@ -1,31 +1,51 @@
 #!/bin/bash
 # Test with:
 #   delivery images to ${IM_SERVER_USER}@${IM_SERVER}. TODO: to registeryserver.
-# set -x
+#'-k' option:
+#   is helpful to skip the image build step for some corner case.
+#   e.g.: the images registry server disk is full.
+
+set -eux
 P_ROOT=$(git rev-parse --show-toplevel)
 COMMON_FILE=${P_ROOT}/pf_common.sh
 source ${COMMON_FILE}
 
-# container: clean
-for c in $(docker ps -aq); do
+echo "container: clean"
+for c in $(docker ps -a | grep $PREFIX | awk '{print $1}'); do
     docker stop $c
     docker rm $c
 done
 
-# images: clean
-docker images | grep -E "$A|$B|$C|$D|$E|none" | awk '{print $3}' | xargs docker rmi
+if [[ $# -gt 0 && "$1" == "-k" ]]; then
+    echo "skip build docker images... "
 
-# saved images tar: clean
-cd $P_ROOT
-for i in ${SAVED[@]}; do
-    cd $P_ROOT && rm -rf $i
-done
-# images: build and save -> tar files
-
-cd $P_ROOT && ./docker_cli build &&
-    for ((i = 0; i < ${#SAVED[@]}; i++)); do
-        docker save -o ${SAVED[$i]} ${IM[$i]}:${TAG[$i]}
+else
+    echo "images: clean"
+    for i in $(docker images | grep -E "$A|$B|$C|$D|$E" | awk '{print $3}'); do
+        if [[ "$i" != "IMAGE" ]]; then
+            echo "removing image $i"
+            docker rmi $i
+        fi
     done
+
+    echo "saved images tar: clean"
+    cd $P_ROOT
+    for i in ${SAVED[@]}; do
+        if [[ -f $i ]]; then
+            cd $P_ROOT && rm -rf $i
+        fi
+    done
+    echo "images: build and save -> tar files"
+    cd $P_ROOT
+    echo "build docker images "
+    cd $P_ROOT
+    ./deploy/nodejs/install_nodejs.sh
+    make managerapp_vm
+fi
+
+for ((i = 0; i < ${#SAVED[@]}; i++)); do
+    docker save -o ${SAVED[$i]} ${IM[$i]}:${TAG[$i]}
+done
 
 # images under HOME_ON_HOP : clean
 echo "rm -rf ${HOME_ON_HOP};  mkdir -p ${HOME_ON_HOP}; ls -l ${HOME_ON_HOP}" |
@@ -43,17 +63,10 @@ done
 # SCP yml and files (under project root) needed to start containers
 tmp_path="./tmp"
 mkdir -p "$tmp_path/"
-cp "$YAML" "$YAML_VM" "$tmp_path/"
+cp "$YAML" "$YAML_VM" "ENV" "$tmp_path/"
 files=(
-    apps/cloudinfra/*
-    apps/portal/*
-    apps/sockstunnel-py/*
     nodes/elasticsearch/files/usr/share/elasticsearch/config/elasticsearch.yml
     nodes/kibana/files/usr/share/kibana/config/*
-    nodes/logstash/files/usr/share/logstash/config/logstash.yml
-    nodes/logstash/files/usr/share/logstash/pipeline/*
-    nodes/logstash/files/usr/share/logstash/template/*
-    nodes/portal/files/etc/nginx/conf.d
 )
 for f in ${files[@]}; do
     base="$tmp_path/${f%/*}"
@@ -61,13 +74,22 @@ for f in ${files[@]}; do
     cp -r "$P_ROOT/$f" "${base}/"
 done
 
+## record the release version
+echo "$(git describe --tags --long --always) " >$tmp_path/$version
 tar -czvf "${FILES_TAR}" -C "$tmp_path" .
+tar -tf "${FILES_TAR}"
+
 echo "delivering ${FILES_TAR}... "
 sshpass -p ${IM_SERVER_PASS} scp -r "${FILES_TAR}" \
     ${IM_SERVER_USER}@${IM_SERVER}:"${HOME_ON_HOP}"/
 
 # tar file: md5sum in local
-md5sum saved*.tar "${FILES_TAR}"
+
+md5sum saved*.tar "${FILES_TAR}" >./${md5sums}
+cat ./${md5sums}
+sshpass -p ${IM_SERVER_PASS} scp -r ./${md5sums} \
+    ${IM_SERVER_USER}@${IM_SERVER}:"${HOME_ON_HOP}"/
+rm ./${md5sums}
 
 # tar file under hop /tmp: md5sum
 echo "md5sum ${HOME_ON_HOP}/saved*.tar ${HOME_ON_HOP}/${FILES_TAR}; ls -lh ${HOME_ON_HOP} " |
@@ -79,3 +101,4 @@ for t in saved*.tar "${FILES_TAR}"; do
 done
 
 rm -rf "$tmp_path"
+echo "Success"
